LayoutTests/platform/chromium/fast/dom/DOMException/prototype-object-expected.txt

This tests the prototype chain of DOMException objects.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS e.toString() is "Error: HierarchyRequestError: DOM Exception 3"
PASS Object.prototype.toString.call(e) is "[object DOMException]"
FAIL Object.prototype.toString.call(e.__proto__) should be [object DOMExceptionPrototype]. Was [object Object].
PASS Object.prototype.toString.call(e.__proto__.__proto__) is "[object Error]"
FAIL e.constructor.toString() should be [object DOMExceptionConstructor]. Was function DOMException() { [native code] }.
PASS e instanceof DOMException is true
PASS e instanceof Error is true
PASS e.constructor is window.DOMException
PASS e.HIERARCHY_REQUEST_ERR is e.constructor.HIERARCHY_REQUEST_ERR
PASS e.HIERARCHY_REQUEST_ERR is 3
PASS e.code is 3
PASS e.name is "HierarchyRequestError"
PASS e.message is "HierarchyRequestError: DOM Exception 3"
PASS successfullyParsed is true

TEST COMPLETE
